iam: move iam_root, iam_alt_root fixtures to iam.py
